+++ tests/Gedmo/Tree/Fixture/Category.php
@@ -107,6 +107,11 @@ class Category implements NodeInterface
     #[ORM\OneToMany(targetEntity: Article::class, mappedBy: 'article')]
     private $comments;
 
+    /**
+     * @var NodeInterface|null
+     */
+    private $sibling;
+
     public function __construct()
     {
         $this->children = new ArrayCollection();
@@ -151,5 +156,15 @@ class Category implements NodeInterface
     public function getLevel(): ?int
     {
         return $this->level;
+    }
+
+    public function setSibling(NodeInterface $node): void
+    {
+        $this->sibling = $node;
+    }
+
+    public function getSibling(): ?NodeInterface
+    {
+        return $this->sibling;
     }
 }